requirements.txt
web3==7.12.0
packaging
eth-account>=0.8.0
python-dotenv
requests
colorama
aiohttp
rich